Remove cv2 and video generation, temporarily
