+++ growbikenet/functions.py
@@ -27,7 +27,7 @@ def validate_parameters(
         export_file_format,
         export_data_slug,
         export_plots,
-        export_video,
+        # export_video,
         allow_edge_overlaps,
         city_boundary_file,
         street_network_file,
@@ -97,8 +97,8 @@ def validate_parameters(
         raise ValueError("export_file_format must be 'geojson' or 'gpkg'")
     if type(export_plots) is not bool:
         raise TypeError("export_plots must be a boolean")
-    if type(export_video) is not bool:
-        raise TypeError("export_video must be a boolean")
+    # if type(export_video) is not bool:
+    #     raise TypeError("export_video must be a boolean")
     if city_boundary_file is not None and type(city_boundary_file) is not str:
         raise TypeError("city_boundary_file must be None or a string")
     if type(city_boundary_file) is str and not os.path.isfile(city_boundary_file):
