+++ growbikenet/growbikenet.py
@@ -31,7 +31,7 @@ from growbikenet.functions import (
     import_network,
     slugify,
 )
-from growbikenet.visualizations import make_video, create_plots
+from growbikenet.visualization import create_plots
 
 
 def growbikenet(
@@ -47,7 +47,7 @@ def growbikenet(
     export_file_format='geojson',
     export_data_slug=None,
     export_plots=False,
-    export_video=False,
+    # export_video=False,
     allow_edge_overlaps=False,
     city_boundary_file=None,
     street_network_file=None,
@@ -100,8 +100,6 @@ def growbikenet(
         If not set to None, the city_name will be slugified and used as the slug in the filename of the data export.
     export_plots : bool, default False
         If set to True, plots are saved to files, overwriting existing ones.
-    export_video : bool, default False
-        If set to True, video is saved to file (only possible if export_plots is set to True), overwriting existing ones.
     allow_edge_overlaps : bool, default False
         If set to False, removes edge overlaps in consecutive growth stages and deletes growth stages that do not add anything new.
     city_boundary_file : str | None, default None
@@ -176,7 +174,6 @@ def growbikenet(
         export_file_format,
         export_data_slug,
         export_plots,
-        export_video,
         allow_edge_overlaps,
         city_boundary_file,
         street_network_file,
@@ -410,7 +407,7 @@ def growbikenet(
     edges_ranked = edges_ranked.astype({'length': int, 'length_cumulative': int})
 
     # Generate export data filename
-    if export_data or export_plots or export_video:
+    if export_data or export_plots:# or export_video:
         os.makedirs("./results/", exist_ok=True)
         if export_data_slug is None:
             city_string = city_name
@@ -455,7 +452,7 @@ def growbikenet(
         progress_bar.update(1)
         progress_bar.close()
 
-    if export_plots or export_video:
+    if export_plots:# or export_video:
         ### Visualize
 
         # Read in file to plot
@@ -480,18 +477,18 @@ def growbikenet(
             ranking,
         )
 
-        if export_video:
-            os.makedirs("./results/plots/ordering_"+ranking+"/video/", exist_ok=True)
-            make_video(img_folder_name="./results/plots/ordering_"+ranking+"/", fps=5)
+        # if export_video:
+        #     os.makedirs("./results/plots/ordering_"+ranking+"/video/", exist_ok=True)
+        #     make_video(img_folder_name="./results/plots/ordering_"+ranking+"/", fps=5)
 
     print("----------------------------------------------╯")
     if export_data:
         print("Data exported to results/")
     if export_plots:
         print("Plots exported to results/plots/")
-    if export_video:
-        print("Video exported to results/plots/")
-    if export_data or export_plots or export_video:
+    # if export_video:
+    #     print("Video exported to results/plots/")
+    if export_data or export_plots:# or export_video:
         print("----------------------------------------------")
 
     endtime = time.time()
